Create synchronous shelldon-command and allow output killing
Synchronous execution of commands is much more predictable than
asynchronous execution. It makes the display of buffers and the
capture of output to the kill ring much easier as well, so I have
added that command here as well. An analogous asynchronous output
killing command will be added later when I figure out how to do it.